Move StringSearchHelper out from stringbase.{h,cpp}.
