+++ searchlib/src/vespa/searchlib/attribute/CMakeLists.txt
@@ -109,5 +109,6 @@ vespa_add_library(searchlib_attribute OBJECT
     sourceselector.cpp
     stringattribute.cpp
     stringbase.cpp
+    string_search_helper.cpp
     DEPENDS
 )
